Go: static code checks for socialapi
added more static checks and fixed their problems
