+++ go/src/socialapi/workers/payment/paymentmodels/customer.go
@@ -111,7 +111,7 @@ func (c *Customer) DeleteSubscriptionsAndItself() error {
 		if subscription.State != SubscriptionStateActive {
 			err := subscription.Delete()
 			if err != nil {
-				fmt.Printf("Deleting user: %s subscription: %s failed: %v\n", c.Username, subscription.Id, err)
+				fmt.Printf("Deleting user: %s subscription: %d failed: %v\n", c.Username, subscription.Id, err)
 			}
 		} else {
 			fmt.Printf("Tried to delete user: %s with active subscription: %v\n", c.Username, subscription.Id)
