+++ scripts/go-nyetcheck.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+if [[ -z "$GOFILES" ]]; then
+    GOFILES=./go/src/socialapi/...
+fi
+
+./go/bin/go-nyet $GOFILES
